fix definition of devices to match the intended semantic
Device drivers (and devices) are meant to represent *specific* types
of devices. In principle, one should only have to do

Robot.devices do
  add Dev::Hokuyo
end

if it has a hokuyo laser scanner. The point being that if one
defines "generic" devices and device drivers, as in

  driver_for "Camera"

It means, in the end, that the driver can connect to *any* type
of camera (which is possible, but no the case for e.g. prosilicacamera::task)